refactor: extract routers package
