+++ api/pybot/main.py
@@ -10,7 +10,8 @@ from fastapi.responses import JSONResponse
 from fastapi.staticfiles import StaticFiles
 from loguru import logger
 
-from pybot.routers import router
+from pybot.routers.chat import router as chat_router
+from pybot.routers.conversations import router as conversation_router
 from pybot.utils import UserIdHeader
 
 
@@ -22,7 +23,8 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(lifespan=lifespan)
 
-app.include_router(router)
+app.include_router(chat_router)
+app.include_router(conversation_router)
 
 
 @app.get("/api/healthz")
