+++ api/pybot/routers/conversations.py
@@ -3,27 +3,12 @@ from pathlib import Path
 from typing import Annotated
 
 import aiofiles
-from fastapi import APIRouter, Depends, UploadFile, WebSocket, WebSocketDisconnect
-from langchain.chains import ConversationChain
-from langchain.llms import BaseLLM, HuggingFaceTextGenInference
+from fastapi import APIRouter, Depends, UploadFile
 from langchain.memory import RedisChatMessageHistory
-from loguru import logger
 
-from pybot.callbacks import (
-    StreamingLLMCallbackHandler,
-    UpdateConversationCallbackHandler,
-)
 from pybot.config import settings
-from pybot.history import CustomRedisChatMessageHistory
-from pybot.memory import FlexConversationBufferWindowMemory
 from pybot.models import Conversation as ORMConversation
-from pybot.prompts.vicuna import (
-    ai_prefix,
-    ai_suffix,
-    human_prefix,
-    human_suffix,
-    prompt,
-)
+from pybot.routers.dependencies import get_message_history
 from pybot.schemas import (
     ChatMessage,
     Conversation,
@@ -33,28 +18,9 @@ from pybot.schemas import (
 from pybot.utils import UserIdHeader
 
 router = APIRouter(
-    prefix="/api",
+    prefix="/api/conversations",
     tags=["conversation"],
 )
-
-
-def get_message_history() -> RedisChatMessageHistory:
-    return CustomRedisChatMessageHistory(
-        url=str(settings.redis_om_url),
-        session_id="sid",  # a fake session id as it is required
-    )
-
-
-def get_llm() -> BaseLLM:
-    return HuggingFaceTextGenInference(
-        inference_server_url=str(settings.inference_server_url),
-        max_new_tokens=1024,
-        temperature=0.8,
-        top_p=0.9,
-        repetition_penalty=1.01,
-        stop_sequences=["</s>"],
-        streaming=True,
-    )
 
 
 @router.get("/conversations", response_model=list[Conversation])
@@ -127,48 +93,3 @@ async def upload_files(
             while content := await file.read(1024):
                 await out_file.write(content)
     return {"filenames": [file.filename for file in files]}
-
-
-@router.websocket("/chat")
-async def generate(
-    websocket: WebSocket,
-    llm: Annotated[BaseLLM, Depends(get_llm)],
-    history: Annotated[RedisChatMessageHistory, Depends(get_message_history)],
-    userid: Annotated[str | None, UserIdHeader()] = None,
-):
-    await websocket.accept()
-    memory = FlexConversationBufferWindowMemory(
-        human_prefix=human_prefix,
-        ai_prefix=ai_prefix,
-        human_suffix=human_suffix,
-        ai_suffix=ai_suffix,
-        memory_key="history",
-        chat_memory=history,
-    )
-    conversation_chain: ConversationChain = ConversationChain(
-        llm=llm,
-        prompt=prompt,
-        verbose=False,
-        memory=memory,
-    )
-
-    while True:
-        try:
-            payload: str = await websocket.receive_text()
-            message = ChatMessage.model_validate_json(payload)
-            history.session_id = f"{userid}:{message.conversation}"
-            streaming_callback = StreamingLLMCallbackHandler(
-                websocket, message.conversation
-            )
-            update_conversation_callback = UpdateConversationCallbackHandler(
-                message.conversation
-            )
-            await conversation_chain.arun(
-                message.content,
-                callbacks=[streaming_callback, update_conversation_callback],
-            )
-        except WebSocketDisconnect:
-            logger.info("websocket disconnected")
-            return
-        except Exception as e:
-            logger.error(f"Something goes wrong, err: {e}")
